less than one day can be ints

+++ src/org/adligo/i/util/client/DateTime.java
@@ -25,11 +25,11 @@ public class DateTime {
 	public static final String FULL_YEAR_TIME_FORMAT = "MM/dd/yyyy hh:mm a SSS";
 	public static final String DEFAULT_DATE_TIME_TIMEZONE_FORMAT = "MM/dd/yyyy hh:mm a SSS Z";
 	
-	public static final long ONE_MINUTE = 60 * 1000;
-	public static final long THREE_MINUTES = 3 * 60 * 1000;
-	public static final long FIVE_MINUTES = 5 * 60 * 1000;
-	public static final long TEN_MINUTES = 10 * 60 * 1000;
-	public static final long ONE_HOUR = 60 * 60 * 1000;
+	public static final int ONE_MINUTE = 60 * 1000;
+	public static final int THREE_MINUTES = 3 * 60 * 1000;
+	public static final int FIVE_MINUTES = 5 * 60 * 1000;
+	public static final int TEN_MINUTES = 10 * 60 * 1000;
+	public static final int ONE_HOUR = 60 * 60 * 1000;
 	public static final long ONE_DAY = 24 * 60 * 60 * 1000;
 	
 	public static final short JANUARY = 1;
